02_backend/requirements.txt
# Helix Backend Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic==2.5.3
python-dotenv==1.0.0
python-multipart==0.0.6

# Security
bcrypt==4.1.2
PyJWT==2.8.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
